+++ src/remote/activitypub/act/delete/index.ts
@@ -1,6 +1,5 @@
-import Resolver from '../resolver';
-import Post from '../../../models/post';
-import { createDb } from '../../../queue';
+import Resolver from '../../resolver';
+import deleteNote from './note';
 
 export default async (actor, activity): Promise<void> => {
 	if ('actor' in activity && actor.account.uri !== activity.actor) {
@@ -15,14 +14,5 @@ export default async (actor, activity): Promise<void> => {
 	case 'Note':
 		deleteNote(object);
 		break;
-	}
-
-	async function deleteNote(note) {
-		const post = await Post.findOneAndDelete({ uri: note.id });
-
-		createDb({
-			type: 'deletePostDependents',
-			id: post._id
-		}).delay(65536).save();
 	}
 };
