+++ src/remote/activitypub/act/delete/note.ts
@@ -1,0 +1,11 @@
+import Post from '../../../../models/post';
+import { createDb } from '../../../../queue';
+
+export default async function(note) {
+	const post = await Post.findOneAndDelete({ uri: note.id });
+
+	createDb({
+		type: 'deletePostDependents',
+		id: post._id
+	}).delay(65536).save();
+}
